requirements.txt
fastapi
uvicorn
sqlalchemy
pg8000
python-multipart
pydantic
pillow
pytesseract
